Set up Codesandbox CI
