+++ .codesandbox/ci.json
@@ -1,0 +1,4 @@
+{
+  "sandboxes": ["/examples"],
+  "node": "14"
+}
