+++ examples/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "examples",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
+  },
+  "dependencies": {
+    "@react-three/drei": "^9.53.0",
+    "@react-three/fiber": "^8.10.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "three": "^0.148.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/three": "0.148.0",
+    "@vitejs/plugin-react": "^3.0.1",
+    "typescript": "^4.9.0",
+    "vite": "^4.0.4"
+  }
+}
