+++ examples/src/index.tsx
@@ -1,64 +1,10 @@
-import * as React from 'react'
-import * as THREE from 'three'
-import * as ReactDOM from 'react-dom/client'
-import { VRButton, XR, Hands, useXR, Interactive, useHitTest, Controllers } from '@react-three/xr'
-import { Box, Text } from '@react-three/drei'
-import { useFrame, Canvas } from '@react-three/fiber'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { App } from "./App";
 
-function Button(props: JSX.IntrinsicElements['mesh']) {
-  const [hover, setHover] = React.useState(false)
-  const [color, setColor] = React.useState(0x123456)
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
 
-  return (
-    <Interactive onSelect={() => setColor((Math.random() * 0xffffff) | 0)} onHover={() => setHover(true)} onBlur={() => setHover(false)}>
-      <Box {...props} args={[0.4, 0.1, 0.1]} scale={hover ? 1.5 : 1}>
-        <meshStandardMaterial color={color} />
-        {false && (
-          <Text position={[0, 0, 0.06]} fontSize={0.05} color="#000" anchorX="center" anchorY="middle">
-            Hello react-xr!
-          </Text>
-        )}
-      </Box>
-    </Interactive>
-  )
-}
-
-function PlayerExample() {
-  const player = useXR((state) => state.player)
-  useFrame(() => void (player.rotation.x = player.rotation.y += 0.01))
-
-  return null
-}
-
-function HitTestExample() {
-  const boxRef = React.useRef<THREE.Mesh>(null!)
-  useHitTest((hitMatrix) => {
-    hitMatrix.decompose(boxRef.current.position, boxRef.current.quaternion, boxRef.current.scale)
-  })
-
-  return <Box ref={boxRef} args={[0.1, 0.1, 0.1]} />
-}
-
-function App() {
-  return (
-    <>
-      <VRButton onError={(e) => console.error(e)} />
-      <Canvas>
-        <XR>
-          <ambientLight intensity={0.5} />
-          <pointLight position={[5, 5, 5]} />
-          <Hands
-          // modelLeft="/hand-left.gltf"
-          // modelRight="/hand-right.gltf"
-          />
-          <Button position={[0, 0.8, -1]} />
-          <Controllers />
-          {false && <PlayerExample />}
-          {false && <HitTestExample />}
-        </XR>
-      </Canvas>
-    </>
-  )
-}
-
-ReactDOM.createRoot(document.getElementById('root')!).render(<App />)
