+++ examples/tsconfig.json
@@ -1,4 +1,29 @@
 {
-  "extends": "../tsconfig",
-  "include": ["src/**/*"]
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "target": "ESNext",
+    "useDefineForClassFields": true,
+    "lib": [
+      "DOM",
+      "DOM.Iterable",
+      "ESNext"
+    ],
+    "allowJs": false,
+    "skipLibCheck": true,
+    "strict": true,
+    "forceConsistentCasingInFileNames": true,
+    "module": "ESNext",
+    "moduleResolution": "Node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "jsx": "react-jsx",
+    "paths": {
+      "@react-three/xr": ["../src"]
+    }
+  },
+  "include": [
+    "src/**/*",
+    "vite.config.ts"
+  ]
 }
